// src/app/components/merchandising/forecast/forecast-file-upload/model/forecast.ts
export class Forecast {
  // forecastId: number;
  // style: string;
  // fgModel: string;
  // fgSizeCode: string;
  // fgSize: string;
  // seasonYearId: number;
  // seasonName: string;
  // dataYearId: number;
  // dataYearName: string;
  // yearName: string;
  // weekId: number;
  // weekName: string;
  // fileUploadDate: moment.Moment;

mstId: number;
dtlId: number;
buyerId: number;
seasonId: number;
yearId: number;
styleId: number;
styleName: string;
companyName: string;
modelCode: string;
modelName: string;
sectorRetail: string;
variant: string;
item: string;
itemName: string;
sizeName: string;
sizeCode: string;
wareHouseName: string;
snapshots: string;
snapshotWeek: number;
snapshotYear: number;
uploadDate: Date;
receivedDate: Date;
status: string;
weekYear: string;
foreCastYearNames: string;
weekNo: number;
yearNo: number;
value: number;



  week1: number;
  week2: number;
  week3: number;
  week4: number;
  week5: number;
  week6: number;
  week7: number;
  week8: number;
  week9: number;
  week10: number;
  week11: number;
  week12: number;
  week13: number;
  week14: number;
  week15: number;
  week16: number;
  week17: number;
  week18: number;
  week19: number;
  week20: number;
  week21: number;
  week22: number;
  week23: number;
  week24: number;
  week25: number;
  week26: number;
  week27: number;
  week28: number;
  week29: number;
  week30: number;
  week31: number;
  week32: number;
  week33: number;
  week34: number;
  week35: number;
  week36: number;
  week37: number;
  week38: number;
  week39: number;
  week40: number;
  week41: number;
  week42: number;
  week43: number;
  week44: number;
  week45: number;
  week46: number;
  week47: number;
  week48: number;
  week49: number;
  week50: number;
  week51: number;
  week52: number;
  week53: number;
  week54: number;
  week55: number;
  week56: number;
  week57: number;
  week58: number;
  week59: number;
  week60: number;
  week61: number;
  week62: number;
  week63: number;
  week64: number;
  week65: number;
  week66: number;
  week67: number;
  week68: number;
  week69: number;
  week70: number;
  week71: number;
  week72: number;
  week73: number;
  week74: number;
  week75: number;
  week76: number;
  week77: number;
  week78: number;
  week79: number;
  week80: number;
  week81: number;
  week82: number;
  week83: number;
  week84: number;
  week85: number;
  week86: number;
  week87: number;
  week88: number;
  week89: number;
  week90: number;
  week91: number;
  week92: number;
  week93: number;
  week94: number;
  week95: number;
  week96: number;
  week97: number;
  week98: number;
  week99: number;
  week100: number;
  week101: number;
  week102: number;
  week103: number;
  week104: number;
  week105: number;
  week106: number;
}